move test using rollback_registry fixture in bloks tests
